+++ packages/solid/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Fix date support for `insert` and `replace` method
+
 ## v0.18.0 (June 23, 2023)
 
 - Rename `autoFocus` in `Field` component to `autofocus` (issue #92)